baseline-refresh: slim formatter further, rescore CRAP/MI/coverage
Move all small pure helpers (deriveStorySymbol, computeProgress,
slugifyHeading, waveHeadingText, renderProgressBar, deriveWaveStatus,
renderWaveSections, topoSortTasks, computeStoryProgress) out of
manifest-formatter.js into a new sibling manifest-helpers.js. The
formatter re-exports each name so external call-sites import paths
stay unchanged.

Also extract renderManifestHeader / renderManifestBody /
renderManifestTelemetry private helpers in the formatter so the
content-hash-memoised _formatManifestMarkdownUncached drops from
CRAP 14.27 → 2.0 (cyclomatic = 2 vs prior 11). manifest-render-waves.js
now imports from manifest-helpers.js, breaking the previous ESM cycle
with the formatter.

Refresh all three baselines (CRAP, MI, coverage) to record the new
file layout plus the per-method drops. The formatter is now 232 LOC
(target <500), buildManifestFromSpec at CRAP 7,
renderNestedWaveSections at CRAP 9, _formatManifestMarkdownUncached
at CRAP 2.
